Experimenting with TabView() and having multiple Views
