+++ TA Signals/TempEmptyView.swift
@@ -1,0 +1,20 @@
+//
+//  TempEmptyView.swift
+//  TA Signals
+//
+//  Created by Gabriel Maldonado Almendra on 27/7/21.
+//
+
+import SwiftUI
+
+struct TempEmptyView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct TempEmptyView_Previews: PreviewProvider {
+    static var previews: some View {
+        TempEmptyView()
+    }
+}
